requirements.txt
openai>=1.0.0
aiohttp>=3.8.0
asyncio
python-dotenv>=1.0.0
# Core dependencies for WhatsApp AI Assistant
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
python-dotenv>=1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
